% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pem.R
\docType{data}
\name{bs4Dash_deps}
\alias{bs4Dash_deps}
\title{Get bs4Dash dependencies}
\format{
An object of class \code{list} of length 10.
}
\usage{
bs4Dash_deps
}
\value{
bs4Dash_deps is a list of html_dependency.
}
\description{
List containing all bs4Dash dependencies.
}
\keyword{datasets}
