% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pem.R
\docType{data}
\name{bs4Dash_ui}
\alias{bs4Dash_ui}
\title{Import bs4Dash dependencies}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 2.
}
\usage{
bs4Dash_ui
}
\value{
bs4Dash_ui is a HTML object.
}
\description{
HMTL object containing all bs4Dash dependencies.
}
\keyword{datasets}
