% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{compute_couloir}
\alias{compute_couloir}
\title{compute_couloir function}
\usage{
compute_couloir(data_raw_name, var_name_id, var_name_age, var_name_perf)
}
\arguments{
\item{data_raw_name}{a character stating the name of the dataset.}

\item{var_name_id}{a character stating the name of the id variable.}

\item{var_name_age}{a character stating the name of the age variable.}

\item{var_name_perf}{a character stating the name of the performance variable.}
}
\value{
compute_couloir() returns a dataset with the performance deciles computed by age.
}
\description{
Function to compute the performance deciles by age.
}
\examples{
# compute_couloir("dt", "licence", "age", "temps")
}
