% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{compute_couloir_SQL}
\alias{compute_couloir_SQL}
\title{compute_couloir_SQL function}
\usage{
compute_couloir_SQL(
  con,
  data_raw_name,
  var_name_id,
  var_name_age,
  var_name_perf,
  nb_couloir
)
}
\arguments{
\item{data_raw_name}{a character stating the name of the dataset.}

\item{var_name_id}{a character stating the name of the id variable.}

\item{var_name_age}{a character stating the name of the age variable.}

\item{var_name_perf}{a character stating the name of the performance variable.}
}
\value{
compute_couloir() returns a dataset with the performance deciles computed by age.
}
\description{
Function to compute the performance deciles by age.
}
\examples{
# compute_couloir("dt", "licence", "age", "temps")
}
