% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pem.R
\name{compute_deriv1}
\alias{compute_deriv1}
\title{Compute the derivative}
\usage{
compute_deriv1(parametre, info, FUN, type = "min", min_age = 10, max_age = 50)
}
\arguments{
\item{parametre}{a data.frame of the parameters estimated by the model.}

\item{info}{a data.frame containing the information of the transformation used on the data to compute the model.}

\item{FUN}{the function to fit.}

\item{type}{("min" by default) indicates whether the function estimates a performance to be maximised ("max") or minimised ("min").}

\item{min_age}{the minimum age value to compute the estimation (10 by default).}

\item{max_age}{the maximum age value to compute the estimation (50 by default).}
}
\value{
compute_deriv1() returns a data.frame.
}
\description{
Function to compute the derivative of the estimation.
}
