% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_age_performance.R
\name{compute_metric}
\alias{compute_metric}
\title{compute_metric function}
\usage{
compute_metric(x, y, para, FUN, metric = "R2")
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}

\item{para}{a vector of parameters.}

\item{FUN}{a function.}

\item{metric}{a character string indicating which metrics to compute. One of "R2" (default), "R2a", "AIC" or "BIC").}
}
\value{
compute_metric() returns the chosen metrics.
}
\description{
Function to compute the R2, R2 adjusted, AIC or BIC.
}
