% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_age_performance.R
\name{compute_r2}
\alias{compute_r2}
\title{compute_r2 function}
\usage{
compute_r2(x, y, para, FUN)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}

\item{para}{a vector of parameters.}

\item{FUN}{a function.}
}
\value{
compute_r2() returns the R2.
}
\description{
Function to compute the R2.
}
