% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_age_performance.R
\name{compute_r2a}
\alias{compute_r2a}
\title{compute_r2a function}
\usage{
compute_r2a(x, y, para, FUN)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}

\item{para}{a vector of the parameters.}

\item{FUN}{a function.}
}
\value{
compute_r2a() returns the R2 adjusted.
}
\description{
Function to compute the R2 adjusted.
}
