% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mils.R
\name{emils}
\alias{emils}
\title{Explore a Markov index load state.}
\usage{
emils(result)
}
\arguments{
\item{result}{a dataframe containing workload and state data.}
}
\value{
emils() returns a shiny app.
}
\description{
Function displaying a shiny app to explore a Markov index load state.
}
