% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_model.R
\name{epem}
\alias{epem}
\title{Explore a performance estimation model}
\usage{
epem(path)
}
\arguments{
\item{path}{the path where to find the SQL database.}
}
\value{
epem() returns a shiny app.
}
\description{
Function displaying a shiny app to explore a performance estimation model.
}
