% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estac}
\alias{estac}
\title{moyenne d’un vecteur
Une fonction pour faire une moyenne en enlevant les valeurs manquantes}
\usage{
estac(x)
}
\arguments{
\item{x}{un vecteur numerique}
}
\value{
la fonction renvoie la moyenne d'un vecteur
}
\description{
moyenne d’un vecteur
Une fonction pour faire une moyenne en enlevant les valeurs manquantes
}
