% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estap}
\alias{estap}
\title{moyenne d’un vecteur
Une fonction pour faire une moyenne en enlevant les valeurs manquantes}
\usage{
estap(
  age,
  performance,
  FUN,
  nbpara,
  niter = 10,
  borne = -Inf,
  initial = runif(nbpara),
  plot = FALSE,
  grid.age = seq(min(age), max(age), 0.1)
)
}
\arguments{
\item{age}{un vecteur numerique}

\item{performance}{un vecteur numerique}

\item{FUN}{un vecteur numerique}

\item{nbpara}{un vecteur numerique}

\item{niter}{un vecteur numerique}

\item{borne}{un vecteur numerique}

\item{initial}{un vecteur numerique}

\item{plot}{un vecteur numerique}

\item{grid.age}{un vecteur numerique}
}
\value{
la fonction renvoie la moyenne d'un vecteur
}
\description{
moyenne d’un vecteur
Une fonction pour faire une moyenne en enlevant les valeurs manquantes
}
