% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{f_imap}
\alias{f_imap}
\title{IMAP function}
\usage{
f_imap(t, p, type = "min")
}
\arguments{
\item{t}{a numeric vector}

\item{p}{a vector of the five parameters}

\item{type}{indicates whether the function estimates a performance to be maximised ("max") or minimised ("min")}
}
\value{
f_imap() returns the evaluation of IMAP
}
\description{
Function to compute IMAP
}
\examples{
f_imap(seq(10,40,0.1), c(10,0.2,1,0.3,80), type = "min")
}
