% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{f_moore}
\alias{f_moore}
\title{The Moore function
Function to compute the Moore equation}
\usage{
f_moore(t, p, type = "min")
}
\arguments{
\item{t}{a numeric vector}

\item{p}{a vector of the four parameters}

\item{type}{indicates whether the function estimates a performance to be maximised ("max") or minimised ("min")}
}
\value{
f_moore() returns the evaluation of the Moore function
}
\description{
The Moore function
Function to compute the Moore equation
}
\examples{
f_moore(seq(10,40,0.1), c(10,0.2,1,0.3), type = "min")
}
