% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_data}
\alias{get_data}
\title{Import data from SQL database}
\usage{
get_data(query, path)
}
\arguments{
\item{query}{a SQL query.}

\item{path}{the path where to find the SQL database.}
}
\value{
get_data() returns a dataframe.
}
\description{
Function to connect to a SQL database and import data.
}
