% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pem.R
\name{my_box}
\alias{my_box}
\title{bs4Dash box}
\usage{
my_box(
  ...,
  title,
  width = 12,
  collapsible = TRUE,
  maximizable = TRUE,
  closable = FALSE
)
}
\arguments{
\item{...}{bs4Dash box arguments.}

\item{title}{a title.}

\item{width}{the width of the box (12 by default).}

\item{collapsible}{boolean. If TRUE, display a button in the upper right that allows the user to collapse the box (TRUE by default).}

\item{maximizable}{boolean. If TRUE, the card can be displayed in full screen mode (TRUE by default).}

\item{closable}{boolean.    If TRUE, display a button in the upper right that allows the user to close the box (FALSE by default).}
}
\value{
my_box() returns a HTML object.
}
\description{
Function creating a bs4Dash box.
}
