% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_age_performance.R
\name{plot_init}
\alias{plot_init}
\title{plot_init function}
\usage{
plot_init(data, FUN, age, para, titre)
}
\arguments{
\item{data}{a dataframe containing age performance data.}

\item{FUN}{a function.}

\item{age}{a vector containing age data.}

\item{para}{a vector of the parameters.}

\item{titre}{the title for the plot.}
}
\value{
plot_init() returns the plot of FUN fitting the data.
}
\description{
Function to plot the fit on data of a function.
}
