% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mils.R
\name{plot_state}
\alias{plot_state}
\title{Plot the workload}
\usage{
plot_state(ath, data, n_state)
}
\arguments{
\item{ath}{athlete id.}

\item{data}{a data.frame containing workload and state data.}

\item{n_state}{the number of state.}
}
\value{
plot_state() returns a ggplot2 object.
}
\description{
Function to plot the workload of an athlete.
}
