% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mils.R
\name{plot_workload}
\alias{plot_workload}
\title{Plot the workload}
\usage{
plot_workload(ath, data)
}
\arguments{
\item{ath}{athlete id.}

\item{data}{a data.frame containing workload data.}
}
\value{
plot_workload() returns a ggplot2 object.
}
\description{
Function to plot the workload of an athlete.
}
