% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pem.R
\name{compute_estimation}
\alias{compute_estimation}
\title{Compute estimation}
\usage{
compute_estimation(reactive, type = "max")
}
\arguments{
\item{reactive}{a list (reactive of the shiny app created by 'epem' function).}

\item{type}{indicates whether performance is to be maximized or minimized ('max' by default).}
}
\value{
compute_estimation() returns a data.frame.
}
\description{
Function to compute the estimation of the performance.
}
