% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pem.R
\name{epem}
\alias{epem}
\title{Explore a performance estimation model}
\usage{
epem(path, type = "max", table_name = "parameter")
}
\arguments{
\item{path}{the path where to find the SQL database.}

\item{type}{indicates whether performance is to be maximized or minimized ('max' by default).}

\item{table_name}{name of the SQL table containing the parameters ('parameter' by default).}
}
\value{
epem() returns a shiny app.
}
\description{
Function displaying a shiny app to explore a performance estimation model.
}
