% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_squares.R
\name{estap}
\alias{estap}
\title{Models age performance relationship}
\usage{
estap(
  age,
  performance,
  FUN,
  nbpara,
  niter = 10,
  borne = -Inf,
  initial = runif(nbpara),
  plot = FALSE,
  grid.age = seq(min(age), max(age), 0.1)
)
}
\arguments{
\item{age}{a vector containing age data.}

\item{performance}{a vector containing performance data.}

\item{FUN}{the function to fit.}

\item{nbpara}{the number of parameters in FUN.}

\item{niter}{number of iterations (10 by default).}

\item{borne}{a vector containing the lower bound of the parameters (-Inf by default).}

\item{initial}{a vector containing initial values of the parameters (runif(nbpara) by default).}

\item{plot}{boolean indicating whether a graph of the result should be displayed (FALSE by default).}

\item{grid.age}{a vector of x.axis values (seq(min(age), max(age), 0.1) by default).}
}
\value{
estap() returns a list including results of the least squares method.
}
\description{
Function performing the least squares method to estimate the relation between age and performance.
}
