% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_squares_method.R
\name{estap}
\alias{estap}
\title{Models age performance relationship}
\usage{
estap(
  age,
  performance,
  FUN,
  nbpara,
  niter = 10,
  borne = -Inf,
  initial = runif(nbpara),
  plot = FALSE,
  grid.age = seq(min(age), max(age), 0.1)
)
}
\arguments{
\item{age}{a vector containing age data}

\item{performance}{a vector containing performance data}

\item{FUN}{the function to fit}

\item{nbpara}{the number of parameters in FUN}

\item{niter}{un vecteur numerique}

\item{borne}{a vector containing the lower bound of the parameters}

\item{initial}{a vector containing initial values of the parameters}

\item{plot}{boolean indicating whether a graph of the result should be displayed0}

\item{grid.age}{a vector of x.axis values}
}
\value{
estap() returns a list including results of the least squares method
}
\description{
Function performing the least squares method to estimate the relation between age and performance.
}
