% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_age_performance.R
\name{f_moore}
\alias{f_moore}
\title{The Moore function}
\usage{
f_moore(t, p, type = "min")
}
\arguments{
\item{t}{a numeric vector.}

\item{p}{a vector of the four parameters.}

\item{type}{("min" by default) indicates whether the function estimates a performance to be maximised ("max") or minimised ("min").}
}
\value{
f_moore() returns the evaluation of the Moore function.
}
\description{
Function to compute the Moore equation.
}
\examples{
age <- seq(10,40,0.1)
parameters <- c(10,0.2,1,0.3)
f_moore(age, parameters, type = "min")
}
