% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_data}
\alias{get_data}
\title{Import data from SQL database}
\usage{
get_data(query, args)
}
\arguments{
\item{query}{a SQL query.}

\item{args}{list of arguments for dbConnect function.}
}
\value{
get_data() returns a dataframe.
}
\description{
Function to connect to a SQL database and import data.
}
