% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pem.R
\name{plot_estimation}
\alias{plot_estimation}
\title{Plot the estimation}
\usage{
plot_estimation(reactive, type = "max")
}
\arguments{
\item{reactive}{a list (reactive of the shiny app created by 'epem' function).}

\item{type}{indicates whether performance is to be maximized or minimized ('max' by default).}
}
\value{
plot_estimation() returns a plotly object.
}
\description{
Function to plot the estimation of the performance.
}
